+++ module/actor/actor-laboratory-sheet.js
@@ -182,6 +182,11 @@ export class ArM5eLaboratoryActorSheet extends ArM5eActorSheet {
           points: result.waste
         }
       )}`;
+      if (context.owner.system.woundsTotal != 0) {
+        context.planning.message += `<br/> ${game.i18n.format("arm5e.lab.planning.msg.wounded", {
+          penalty: context.owner.system.woundsTotal
+        })}`;
+      }
     }
     context.planning.duration = result.duration;